Make sure the chunk size matches cloudvolume setup
Cloudvolume refuses to write partial chunks unless at the boundary of
the volume, make sure the dimension of our atomic chunk are multiples of
that of the cloudvolume chunk

+++ dags/sanitycheck_dag.py
@@ -191,6 +191,10 @@ def check_cv_data():
         Variable.set("param", param, serialize_json=True)
         slack_message(":exclamation:*Process dataset in 512x512x64 chunks by default*")
 
+    if any( x%y != 0 for x, y in zip(param["CHUNK_SIZE"], param["CV_CHUNK_SIZE"]) ):
+        slack_message(":u7981:*ERROR: CHUNK_SIZE must be multiples of CV_CHUNK_SIZE in each dimension: {} vs {}*".format(param["CHUNK_SIZE"], param["CV_CHUNK_SIZE"]))
+        raise ValueError('CHUNK_SIZE must be multiples of CV_CHUNK_SIZE')
+
     for k in mount_secrets:
         os.remove(os.path.join(cv_secrets_path, k))
